+++ src/app/presentation/components/workspace-components/workspace-info/workspace-info.component.html
@@ -1,3 +1,6 @@
-<div class="right-side-title">
-  <span> Workspace info </span>
+<div>
+  <span>
+    <button [disabled]= isOwner
+      mat-raised-button class="warning" (click)="leave()">Leave</button>
+  </span>
 </div>
