+++ src/app/presentation/components/workspace-components/workspace-settings/workspace-settings.component.html
@@ -3,9 +3,3 @@
 </div>
 <br><div><button mat-raised-button color="Primary" (click)='modalUpdate()'>Update</button></div>
 <br>
-<div>
-  <span>
-    <button [disabled]="isOwner" 
-      mat-raised-button class="warning" (click)="leave()">Leave</button>
-  </span>
-</div>
