+++ src/app/presentation/components/workspace-components/workspace-settings/workspace-settings.component.ts
@@ -1,13 +1,10 @@
 import { WorkspaceInfo } from '../../../../core/models/workspace/workspaceInfo';
 import { Subscription } from 'rxjs';
 import { MatDialog } from '@angular/material/dialog';
-import { ActivatedRoute, Router } from '@angular/router';
+import { ActivatedRoute } from '@angular/router';
 import { Component, OnInit } from '@angular/core';
 import { WorkspaceInviteComponent } from '../workspace-invite/workspace-invite.component';
 import { WorkspaceUpdateComponent } from '../workspace-update/workspace-update.component';
-import { AlertService } from 'src/app/core/services/alerts.service';
-import { WorkspaceService } from 'src/app/core/services/workspace.service';
-import { UserService } from 'src/app/core/services/user.service';
 
 @Component({
   selector: 'app-workspace-settings',
@@ -19,13 +16,8 @@ export class WorkspaceSettingsComponent implements OnInit {
   protected routeSub: Subscription;
   workspaceId: number;
   workspace: WorkspaceInfo;
-  isOwner: boolean;
 
-  constructor(private route: ActivatedRoute, 
-    public dialog: MatDialog,
-    private router: Router, 
-    private alertService: AlertService,
-    private workspaceServise: WorkspaceService,) {}
+  constructor(private route: ActivatedRoute, public dialog: MatDialog,) {}
 
   ngOnInit() {
     this.route.parent?.params.subscribe(
@@ -33,14 +25,6 @@ export class WorkspaceSettingsComponent implements OnInit {
       {
         this.workspaceId = Number(params['id']);
        });
-       this.workspaceServise.getWorkspaceInfo(this.workspaceId)
-       .subscribe((data: WorkspaceInfo) => {
-        this.workspace = data;
-        
-        if (this.workspace.role==1) {
-            this.isOwner = true
-          }
-      });
   }
 
   modalInvites() {
@@ -56,17 +40,4 @@ export class WorkspaceSettingsComponent implements OnInit {
         dialogRef.close();
     });
   }
-
-  async leave (){
-      if  (await this.alertService.confirmMessage('You will be deleted from all workspace tasks!', 
-        'Are you sure?', 
-        'Yes, leave!'))
-      {
-      this.alertService.successMessage();
-      this.workspaceServise.leaveFromWorkspace(this.workspaceId).subscribe(
-        () => { 
-          this.router.navigate(['/user/workspaces']);
-        })
-      }
-    }
 }
